backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.27
alembic>=1.7.1
psycopg2-binary>=2.9.1
python-dotenv==1.0.1

# AI and ML dependencies
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.0
torchvision>=0.16.0
monai==1.3.1
numpy==1.26.4
pandas>=1.3.0
scikit-learn>=0.24.2
opencv-python==4.9.0.80
pydicom==2.4.4
pillow

# PDF and reporting
reportlab==4.1.0
weasyprint==60.2

# Testing and development
pytest==8.0.1
black>=21.7b0
flake8>=3.9.2
mypy>=0.910

# Utilities
tqdm>=4.62.2
requests>=2.26.0
python-magic>=0.4.24
httpx==0.26.0 